+++ bindings/cpu_riscv64.h
@@ -1,0 +1,50 @@
+/*
+ * Copyright (c) 2015-2026 Contributors as noted in the AUTHORS file
+ *
+ * This file is part of Solo5, a sandboxed execution environment.
+ *
+ * Permission to use, copy, modify, and/or distribute this software
+ * for any purpose with or without fee is hereby granted, provided
+ * that the above copyright notice and this permission notice appear
+ * in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL
+ * WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE
+ * AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR
+ * CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS
+ * OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT,
+ * NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
+ * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+#ifndef __CPU_RISCV64_H__
+#define __CPU_RISCV64_H__
+
+/* memory defines */
+#define PAGE_SIZE  4096
+#define PAGE_SHIFT 12
+#define PAGE_MASK  ~(0xfff)
+
+#ifndef ASM_FILE
+
+/*
+ * The remainder of this file is used only from C.
+ */
+static inline uint64_t cpu_cntvct(void)
+{
+    uint64_t val;
+
+    /* rdtime and not rdcycle, which Linux leaves trapping in U-mode. */
+    __asm__ __volatile__("rdtime %0" : "=r"(val));
+
+    return val;
+}
+
+static inline void cpu_set_tls_base(uint64_t base)
+{
+    __asm__ __volatile__("mv tp, %0" : : "r"(base));
+}
+
+#endif /* !ASM_FILE */
+
+#endif /* __CPU_RISCV64_H__ */